requirements.txt
spacy
torch
tqdm
streamlit>=1.32.0
pymupdf>=1.23.0
sentence-transformers>=2.6.0
faiss-cpu>=1.8.0
rank-bm25>=0.2.2
nltk>=3.8.1
numpy>=1.26.0
requests>=2.31.0
rouge-score
